+++ e2e/react-native.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  dependencies: {
+    'react-native-google-cast': {
+      platforms: {
+        ios: null, // this will disable autolinking for this package on iOS
+      },
+    },
+  },
+};